// src/features/rps/enums.ts
export enum BetType {
    ROCK = "ROCK",
    PAPER = "PAPER",
    SCISSORS = "SCISSORS",
}

export enum GameStage {
    SELECTING_BETS = "SELECTING_BETS",
    CALCULATING_RESULTS = "CALCULATING_RESULTS",
    SHOWING_RESULTS = "SHOWING_RESULTS",
}

export enum MatchResult {
    WIN = "WIN",
    TIE = "TIE",
    LOSE = "LOSE",
}